notebooks/04-google-adk-agents/requirements.txt
# Google ADK and dependencies
google-adk>=0.1.0
google-genai>=0.1.0

# MCP SDK
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Async support
asyncio>=3.4.3